[FIX] name of messages folder
