--- example/src/components/NewMessages/styles.tsx
@@ -1,21 +1,0 @@
-
-import styled from 'styled-components/native'
-import { colors } from '../../theme'
-
-export const Wrapper = styled.View`
-  flex: 1;
-  background-color: ${colors.white};
-  padding: 0
-`
-export const Header = styled.View`
-  flex-direction: row;
-  align-items: center;
-  border-bottom-width: 1px;
-  border-bottom-color: #d9d9d9;
-  padding-bottom: 10px;
-  padding-horizontal: 20px;
-`
-
-export const TitleHeader = styled.View`
-
-`
